PR for 3887 - move click from col header to sort element (#3891)
* Change click element

Move the eventlistner for the click from 'colEl' to 'arrowEl'. Remove the hover style from the column title and add hover style to the sort arrow using an arrow hover variable.

* Switched to an option headerSortClickElement

This option can be added to the configuration, defaults to 'header'. A css class is added to the correct element to help with styling based on the set option

* Minor tweaks

Change from if conditions to a switch